+++ client/src/lib/forms/disable-save.js
@@ -102,6 +102,7 @@ function disableSave(field, stateData, componentProps, craig) {
     "power_shared_processor_pools",
     "secrets_groups",
     "cloud_logs",
+    "certificates",
   ];
   let isPowerSshKey = field === "ssh_keys" && componentProps.arrayParentName;
   if (contains(stateDisableSaveComponents, field) || isPowerSshKey) {
@@ -166,8 +167,8 @@ function disableSave(field, stateData, componentProps, craig) {
         ? componentProps.craig.vpn_gateways.connections
         : field === "classic_sg_rules"
         ? componentProps.craig.classic_security_groups.classic_sg_rules
-        : field === "secrets_groups"
-        ? componentProps.craig.secrets_manager.secrets_groups
+        : contains(["secrets_groups", "certificates"], field)
+        ? componentProps.craig.secrets_manager[field]
         : componentProps.craig[field]
     ).shouldDisableSave(stateData, componentProps);
   } else return false;