+++ client/src/lib/forms/props-match-state.js
@@ -102,6 +102,7 @@ function propsMatchState(field, stateData, componentProps) {
     // this is to catch in cloud services form to prevent form from crashing the page
     // somewhere a function gets sent on load to deepEqual for propsMatchState. I believe
     // this is happening before component render
+
     return true;
   }
 }